refactor: create_meta
