+++ tests/create_meta/model_card_gold_without_frontmatter.md
@@ -509,7 +509,6 @@ model-index:
     task:
       type: STS
 ---
-
 # all-MiniLM-L6-v2
 This is a [sentence-transformers](https://www.SBERT.net) model: It maps sentences & paragraphs to a 384 dimensional dense vector space and can be used for tasks like clustering or semantic search.
